test(main_loop): happy path of `handle_self_guessed_block`
In commit 1d145d38065fe7d6c173cc2f6b00cc2e870e9b37, the handling of a
new block by `main_loop` is factored out to a separate function, since
this logic can now be invoked in two cases: If the node's own guesser
finds a solution to the PoW-puzzle, and if an external program (like a
GPU guesser) finds a solution to the PoW-puzzle. Previously, this logic
could only be invoked by the node's own guesser. But this factored out
function was not tested. This commit fixes that, at least for the happy
path of this function.